change type of _XMLRPC_STRUCT_MEMBER_OFFSET from unsigned long to size_t
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1478 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/include/xmlrpc-c/util.h
@@ -40,7 +40,7 @@ extern "C" {
 */
 
 #define _XMLRPC_STRUCT_MEMBER_OFFSET(TYPE, MBRNAME) \
-  ((unsigned long)(char*)&((TYPE *)0)->MBRNAME)
+  ((size_t)(char*)&((TYPE *)0)->MBRNAME)
 #define _XMLRPC_STRUCT_MEMBER_SIZE(TYPE, MBRNAME) \
   sizeof(((TYPE *)0)->MBRNAME)
 #define XMLRPC_STRUCTSIZE(TYPE, MBRNAME) \
